requirements.txt
pandas==2.2.3
numpy==1.26.4
Flask==3.1.0
hdbscan==0.8.40
umap-learn==0.5.7
nltk==3.9.1
sentence-transformers==3.3.1
openai==1.86.0
python-dotenv==1.1.0
langchain==0.3.26
langchain-core==0.3.66
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langsmith==0.4.1
scikit-learn
matplotlib
langgraph
langchain-docling